+++ src/config.php
@@ -1,0 +1,28 @@
+<?php
+/**
+ * @copyright Copyright (c) PutYourLightsOn
+ */
+
+/**
+ * Recipe config.php
+ *
+ * This file exists only as a template for the Recipe settings.
+ * It does nothing on its own.
+ *
+ * Don't edit this file, instead copy it to 'craft/config' as 'recipe.php'
+ * and make your changes there to override default settings.
+ *
+ * Once copied to 'craft/config', this file will be multi-environment aware as
+ * well, so you can have different settings groups for each environment, just as
+ * you do for 'general.php'
+ */
+
+return [
+    '*' => [
+        // An application ID for the Edamam Nutrition Analysis API (https://developer.edamam.com/edamam-nutrition-api).
+        //'apiApplicationId' => '1a2b3c4e',
+
+        // An application key for the Edamam Nutrition Analysis API (https://developer.edamam.com/edamam-nutrition-api).
+        //'apiApplicationId' => '1a2b3c4e5f6g7h8i9j0k',
+    ]
+];