+++ src/services/NutritionApi.php
@@ -1,0 +1,90 @@
+<?php
+/**
+ * Recipe plugin for Craft CMS 3.x
+ *
+ * A comprehensive recipe FieldType for Craft CMS that includes metric/imperial
+ * conversion, portion calculation, and JSON-LD microdata support
+ *
+ * @link      https://nystudio107.com
+ * @copyright Copyright (c) 2017 nystudio107
+ */
+
+namespace nystudio107\recipe\services;
+
+use Craft;
+use craft\base\Component;
+use Exception;
+use nystudio107\recipe\Recipe;
+
+/**
+ * @author    nystudio107
+ * @package   Recipe
+ * @since     1.3.0
+ */
+class NutritionApi extends Component
+{
+    /**
+     * Returns nutritional information about a recipe.
+     *
+     * @param array $ingredients
+     * @param int|null $serves
+     * @return array
+     */
+    public function getNutritionalInfo(array $ingredients, int $serves = null): array
+    {
+        if (Recipe::$plugin->settings->hasApiCredentials() === false) {
+            return [];
+        }
+
+        $url = 'https://api.edamam.com/api/nutrition-details'
+            .'?app_id='.Craft::parseEnv(Recipe::$plugin->settings->apiApplicationId)
+            .'&app_key='.Craft::parseEnv(Recipe::$plugin->settings->apiApplicationKey);
+
+        $data = [
+            'ingr' => $ingredients,
+        ];
+
+        if ($serves) {
+            $data['yield'] = $serves;
+        }
+
+        try {
+            $response = Craft::createGuzzleClient()->post($url, ['json' => $data]);
+
+            $result = json_decode($response->getBody());
+
+            $yield = $result->yield ?: 1;
+
+            return [
+                'servingSize' => round($result->totalWeight / $yield, 0).' grams',
+                'calories' => round($result->totalNutrients->ENERC_KCAL->quantity / $yield, 0),
+                'carbohydrateContent' => round($result->totalNutrients->CHOCDF->quantity / $yield, 1),
+                'cholesterolContent' => round($result->totalNutrients->CHOLE->quantity / $yield, 1),
+                'fatContent' => round($result->totalNutrients->FAT->quantity / $yield, 1),
+                'fiberContent' => round($result->totalNutrients->FIBTG->quantity / $yield, 1),
+                'proteinContent' => round($result->totalNutrients->PROCNT->quantity / $yield, 1),
+                'saturatedFatContent' => round($result->totalNutrients->FASAT->quantity / $yield, 1),
+                'sodiumContent' => round($result->totalNutrients->NA->quantity / $yield, 1),
+                'sugarContent' => round($result->totalNutrients->SUGAR->quantity / $yield, 1),
+                'transFatContent' => round($result->totalNutrients->FATRN->quantity / $yield, 1),
+                'unsaturatedFatContent' => round(($result->totalNutrients->FAMS->quantity + $result->totalNutrients->FAPU->quantity) / $yield, 1),
+            ];
+        }
+        catch (Exception $exception) {
+            $message = 'Error fetching nutritional information from API. ';
+
+            switch ($exception->getCode()) {
+                case 401:
+                    $message .= 'Please verify your API credentials.';
+                    break;
+                case 555:
+                    $message .= 'One or more ingredients could not be recognized.';
+                    break;
+            }
+
+            Craft::error($message.$exception->getMessage(), __METHOD__);
+
+            return ['error' => $message];
+        }
+    }
+}